emul-trace-mem: split instructions over subcycles

+++ ceno_emul/src/addr.rs
@@ -21,7 +21,7 @@ pub const WORD_SIZE: usize = 4;
 // Type aliases to clarify the code without wrapper types.
 pub type Word = u32;
 pub type Addr = u32;
-pub type Cycle = u32;
+pub type Cycle = u64;
 pub type RegIdx = usize;
 
 #[derive(Clone, Copy, Default, PartialEq, Eq, Hash)]
